requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
telethon>=1.33.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
qrcode[pil]>=7.4.0
cryptography>=41.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
pillow>=10.0.0
